backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
PyYAML==6.0.2
pypdf==5.9.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.3.0
httpx==0.28.1
openpyxl==3.1.5
